Move views into separate folder for findability
