+++ packages/flutter/apps/experiments/magicians/lib/screens/main/views/camera_view.dart
